requirements.txt
certifi==2018.10.15
dj-database-url==0.5.0
dj-static==0.0.6
Django==2.1.4
django-filter==2.0.0
django-toolbelt==0.0.1
djangorestframework==3.9.0
gunicorn==19.9.0
Markdown==3.0.1
mysqlclient==1.3.14
python-dotenv==0.9.1
pytz==2018.7
static3==0.7.0
